requirements.txt
fastapi[standard]==0.110.0
pydantic==2.7.1
nltk==3.8.1
groq==0.9.0
python-dotenv==1.0.1
httpx==0.27.0
sqlalchemy==2.0.29